third_party/WebKit/LayoutTests/inspector-protocol/debugger/domdebugger-getEventListeners-expected.txt

Event listeners of window
  type:click
  useCapture:true
  lineNumber:3
  columnNumber:33
  handler.type:function
  handler.className:Function
  handler.description:function clickHandler(event) { console.log('click - button - bubbling (registered before attribute)'); }

  type:hover
  useCapture:true
  lineNumber:5
  columnNumber:66
  handler.type:function
  handler.className:Function
  handler.description:function hoverHandler(event) { console.log("hover - button - bubbling"); }

  type:load
  useCapture:false
  lineNumber:164
  columnNumber:18
  handler.type:function
  handler.className:Function
  handler.description:function onload(event) {  runTest();}

  type:scroll
  useCapture:false
  lineNumber:1
  columnNumber:54
  handler.type:function
  handler.className:Function
  handler.description:function (){ consol.log(42) }

  type:scroll
  useCapture:false
  lineNumber:2
  columnNumber:54
  handler.type:function
  handler.className:Function
  handler.description:function (){ consol.log(42) }


Event listeners of div#listeners1
  type:click
  useCapture:true
  lineNumber:2
  columnNumber:33
  handler.type:function
  handler.className:Function
  handler.description:function clickHandler(event) { console.log('click - button - bubbling (registered before attribute)'); }

  type:hover
  useCapture:true
  lineNumber:4
  columnNumber:63
  handler.type:function
  handler.className:Function
  handler.description:function hoverHandler(event) { console.log("hover - button - bubbling"); }

  type:load
  useCapture:false
  lineNumber:162
  columnNumber:33
  handler.type:function
  handler.className:Function
  handler.description:function onload(event) {  return 42;}


Event listeners of div#listeners2
